TP_5/requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
matplotlib==3.7.2